prettier used and file extensions corrected
