+++ public/media-hunt-react-app/src/API/configuration.js
@@ -1,0 +1,9 @@
+const configuration = {
+  get: {
+    configuration: {
+      url: "/configuration",
+    },
+  },
+};
+
+export default configuration;
