--- public/media-hunt-react-app/src/API/configuration.jsx
@@ -1,9 +1,0 @@
-const configuration = {
-    get: {
-      configuration: {
-        url: "/configuration",
-      },
-    },
-  };
-
-  export default configuration;
