+++ public/media-hunt-react-app/src/API/index.js
@@ -3,9 +3,4 @@ import movie from "./movie";
 import tv from "./tv";
 import search from "./search";
 
-export{
-    configuration,
-    movie,
-    tv,
-    search
-}
+export { configuration, movie, tv, search };
