+++ public/media-hunt-react-app/src/API/movie.js
@@ -1,0 +1,24 @@
+const movie = {
+  get: {
+    discover: {
+      url: "/discover/movie",
+    },
+    movie: {
+      url: (id) => `/movie/${id}`,
+    },
+    genres: {
+      url: "/genre/movie/list",
+    },
+    topRated: {
+      url: "/movie/top_rated",
+    },
+    reviews: {
+      url: (id) => `/movie/${id}/reviews`,
+    },
+    cast: {
+      url: (id) => `/movie/${id}/credits`,
+    },
+  },
+};
+
+export default movie;
