--- public/media-hunt-react-app/src/API/movie.jsx
@@ -1,24 +1,0 @@
-const movie = {
-    get: {
-      discover: {
-        url: "/discover/movie",
-      },
-      movie: {
-        url: (id) => `/movie/${id}`,
-      },
-      genres: {
-        url: "/genre/movie/list",
-      },
-      topRated: {
-        url: "/movie/top_rated",
-      },
-      reviews: {
-        url: (id) => `/movie/${id}/reviews`,
-      },
-      cast: {
-        url: (id) => `/movie/${id}/credits`,
-      }
-    },
-  };
-
-  export default movie;
