+++ public/media-hunt-react-app/src/API/search.js
@@ -1,0 +1,12 @@
+const search = {
+  get: {
+    searchMovie: {
+      url: "/search/movie",
+    },
+    searchTv: {
+      url: "/search/tv",
+    },
+  },
+};
+
+export default search;
