--- public/media-hunt-react-app/src/API/search.jsx
@@ -1,15 +1,0 @@
-const search = {
-    get: {
-      multiSearch: {
-        url: "/search/multi",
-      },
-      searchMovie: {
-          url: "/search/movie",
-      },
-      searchTv: {
-          url: "/search/tv",
-      }
-    },
-  };
-
-  export default search;
